test(TransitLandClient): Extract withMockServer
