+++ backend/src/test/kotlin/com/mobilispect/backend/util/MockWebServerUtils.kt
@@ -1,0 +1,12 @@
+package com.mobilispect.backend.util
+
+import okhttp3.mockwebserver.Dispatcher
+import okhttp3.mockwebserver.MockWebServer
+
+fun withMockServer(dispatcher: Dispatcher, block: (MockWebServer) -> Unit) {
+    val mockServer = MockWebServer()
+    mockServer.dispatcher = dispatcher
+    mockServer.start()
+    block(mockServer)
+    mockServer.shutdown()
+}
